+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `systemd-nspawn` and some packages build utilities in restricted sandboxes.
 - Add AppArmor profile for `fatbuildrd` to run `mkosi` in unconfined sandboxes
   without transition to default `unprivileged_userns` profile.
+- conf: Add `img_create_use_sysusersd` boolean parameter in `[format:*]` section
+  to control if _fatbuildr_ system user is created with `sysusers.d`
+  configuration file during image creation or with raw seeds of `/etc/passwd`,
+  `/etc/group` and `/etc/gshadow`.
 - pkgs: Install and enable AppArmor profile for `u-nspawn` on systems that
   support this LSM.